エラー用hooksの場所変更
